Added load_utilities.hpp and moved function implementations out of load_utilities.h
