+++ metaroom_xml_parser/include/load_utilities.h
@@ -6,58 +6,13 @@ namespace semantic_map_load_utilties
 
     /********************************************** MERGED CLOUD UTILITIES ****************************************************************************************/
     template <class PointType>
-    boost::shared_ptr<pcl::PointCloud<PointType>> loadMergedCloudFromSingleSweep(std::string sweepXmlPath, bool verbose=false)
-    {
-        auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweepXmlPath, std::vector<std::string>{"RoomCompleteCloud"},verbose);
-
-        return sweep.completeRoomCloud;
-    }
+    boost::shared_ptr<pcl::PointCloud<PointType>> loadMergedCloudFromSingleSweep(std::string sweepXmlPath, bool verbose=false);
 
     template <class PointType>
-    std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>> loadMergedCloudFromMultipleSweeps(std::string folderPath, bool verbose=false)
-    {
-        std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>> toRet;
-
-        SimpleSummaryParser summary_parser;
-        summary_parser.createSummaryXML(folderPath);
-        auto sweep_xmls = summary_parser.getRooms();
-
-        for (size_t i=0; i<sweep_xmls.size(); i++)
-        {
-            auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweep_xmls[i].roomXmlFile, std::vector<std::string>{"RoomCompleteCloud"}, verbose);
-            toRet.push_back(sweep.completeRoomCloud);
-        }
-
-        return toRet;
-    }
+    std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>> loadMergedCloudFromMultipleSweeps(std::string folderPath, bool verbose=false);
 
     template <class PointType>
-    std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>> loadMergedCloudForTopologicalWaypoint(std::string folderPath, std::string waypoint,bool verbose=false)
-    {
-        std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>> toRet;
-
-        SimpleSummaryParser summary_parser;
-        summary_parser.createSummaryXML(folderPath);
-        auto sweep_xmls = summary_parser.getRooms();
-
-        // first construct waypoint id to sweep xml map
-        std::map<std::string, std::vector<std::string>> waypointToSweepsMap;
-        for (size_t i=0; i<sweep_xmls.size(); i++)
-        {
-            auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweep_xmls[i].roomXmlFile, std::vector<std::string>(), verbose);
-            waypointToSweepsMap[sweep.roomWaypointId].push_back(sweep_xmls[i].roomXmlFile);
-        }
-
-        auto sweepsAtWaypoint = waypointToSweepsMap[waypoint];
-
-        for (size_t i=0; i<sweepsAtWaypoint.size(); i++)
-        {
-            auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweepsAtWaypoint[i], std::vector<std::string>{"RoomCompleteCloud"}, verbose);
-            toRet.push_back(sweep.completeRoomCloud);
-        }
-
-        return toRet;
-    }
+    std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>> loadMergedCloudForTopologicalWaypoint(std::string folderPath, std::string waypoint,bool verbose=false);
 
     /********************************************** INTERMEDIATE CLOUD UTILITIES ****************************************************************************************/
 
@@ -72,173 +27,34 @@ namespace semantic_map_load_utilties
     };
 
     template <class PointType>
-    std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>> loadIntermediateCloudsFromSingleSweep(std::string sweepXmlPath, bool verbose=false)
-    {
-        auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweepXmlPath, std::vector<std::string>{"RoomIntermediateCloud"},verbose);
-
-        return sweep.vIntermediateRoomClouds;
-    }
+    std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>> loadIntermediateCloudsFromSingleSweep(std::string sweepXmlPath, bool verbose=false);
 
     template <class PointType>
-    IntermediateCloudCompleteData<PointType> loadIntermediateCloudsCompleteDataFromSingleSweep(std::string sweepXmlPath, bool verbose=false)
-    {
-        auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweepXmlPath, std::vector<std::string>{"RoomIntermediateCloud"},verbose);
-
-        IntermediateCloudCompleteData<PointType> toRet;
-        toRet.vIntermediateRoomClouds = sweep.vIntermediateRoomClouds;
-        toRet.vIntermediateRoomCloudTransforms = sweep.vIntermediateRoomCloudTransforms;
-        toRet.vIntermediateRoomCloudCamParams = sweep.vIntermediateRoomCloudCamParams;
-        toRet.vIntermediateRGBImages = sweep.vIntermediateRGBImages;
-        toRet.vIntermediateDepthImages = sweep.vIntermediateDepthImages;
-
-        return toRet;
-    }
+    IntermediateCloudCompleteData<PointType> loadIntermediateCloudsCompleteDataFromSingleSweep(std::string sweepXmlPath, bool verbose=false);
 
     template <class PointType>
-    std::vector<std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>>> loadIntermediateCloudsFromMultipleSweeps(std::string folderPath, bool verbose=false)
-    {
-        std::vector<std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>>> toRet;
-
-        SimpleSummaryParser summary_parser;
-        summary_parser.createSummaryXML(folderPath);
-        auto sweep_xmls = summary_parser.getRooms();
-
-        for (size_t i=0; i<sweep_xmls.size(); i++)
-        {
-            auto intClouds = loadIntermediateCloudsFromSingleSweep<PointType>(sweep_xmls[i],verbose);
-            toRet.push_back(intClouds);
-        }
-
-        return toRet;
-    }
+    std::vector<std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>>> loadIntermediateCloudsFromMultipleSweeps(std::string folderPath, bool verbose=false);
 
     template <class PointType>
-    std::vector<IntermediateCloudCompleteData<PointType>>  loadIntermediateCloudsCompleteDataFromMultipleSweeps(std::string folderPath, bool verbose=false)
-    {
-        std::vector<IntermediateCloudCompleteData<PointType>> toRet;
-
-        SimpleSummaryParser summary_parser;
-        summary_parser.createSummaryXML(folderPath);
-        auto sweep_xmls = summary_parser.getRooms();
-
-        for (size_t i=0; i<sweep_xmls.size(); i++)
-        {
-            auto intCloudsComplete = loadIntermediateCloudsCompleteDataFromSingleSweep<PointType>(sweep_xmls[i],verbose);
-            toRet.push_back(intCloudsComplete);
-        }
-
-        return toRet;
-    }
+    std::vector<IntermediateCloudCompleteData<PointType>>  loadIntermediateCloudsCompleteDataFromMultipleSweeps(std::string folderPath, bool verbose=false);
 
     template <class PointType>
-    std::vector<std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>>> loadIntermediateCloudsForTopologicalWaypoint(std::string folderPath, std::string waypoint,bool verbose=false)
-    {
-        std::vector<std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>>> toRet;
+    std::vector<std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>>> loadIntermediateCloudsForTopologicalWaypoint(std::string folderPath, std::string waypoint,bool verbose=false);
 
-        SimpleSummaryParser summary_parser;
-        summary_parser.createSummaryXML(folderPath);
-        auto sweep_xmls = summary_parser.getRooms();
-
-        // first construct waypoint id to sweep xml map
-        std::map<std::string, std::vector<std::string>> waypointToSweepsMap;
-        for (size_t i=0; i<sweep_xmls.size(); i++)
-        {
-            auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweep_xmls[i].roomXmlFile, std::vector<std::string>(), verbose);
-            waypointToSweepsMap[sweep.roomWaypointId].push_back(sweep_xmls[i].roomXmlFile);
-        }
-
-        auto sweepsAtWaypoint = waypointToSweepsMap[waypoint];
-
-        for (size_t i=0; i<sweepsAtWaypoint.size(); i++)
-        {
-            auto intClouds = loadIntermediateCloudsFromSingleSweep<PointType>(sweepsAtWaypoint[i],verbose);
-            toRet.push_back(intClouds);
-        }
-
-        return toRet;
-    }
-
-        template <class PointType>
-        std::vector<IntermediateCloudCompleteData<PointType>> loadIntermediateCloudsCompleteDataForTopologicalWaypoint(std::string folderPath, std::string waypoint,bool verbose=false)
-    {
-        std::vector<IntermediateCloudCompleteData<PointType>> toRet;
-
-        SimpleSummaryParser summary_parser;
-        summary_parser.createSummaryXML(folderPath);
-        auto sweep_xmls = summary_parser.getRooms();
-
-        // first construct waypoint id to sweep xml map
-        std::map<std::string, std::vector<std::string>> waypointToSweepsMap;
-        for (size_t i=0; i<sweep_xmls.size(); i++)
-        {
-            auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweep_xmls[i].roomXmlFile, std::vector<std::string>(), verbose);
-            waypointToSweepsMap[sweep.roomWaypointId].push_back(sweep_xmls[i].roomXmlFile);
-        }
-
-        auto sweepsAtWaypoint = waypointToSweepsMap[waypoint];
-
-        for (size_t i=0; i<sweepsAtWaypoint.size(); i++)
-        {
-            auto intCloudsComplete = loadIntermediateCloudsCompleteDataFromSingleSweep<PointType>(sweepsAtWaypoint[i],verbose);
-            toRet.push_back(intCloudsComplete);
-        }
-
-        return toRet;
-    }
+    template <class PointType>
+    std::vector<IntermediateCloudCompleteData<PointType>> loadIntermediateCloudsCompleteDataForTopologicalWaypoint(std::string folderPath, std::string waypoint,bool verbose=false);
 
     /********************************************** INTERMEDIATE POSITION IMAGES UTILITIES ****************************************************************************************/
-    template <class PointType>
-    std::vector<typename SimpleXMLParser<PointType>::IntermediatePositionImages> loadIntermediatePositionImagesFromSingleSweep(std::string sweepXmlPath, bool verbose=false)
-    {
-        auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweepXmlPath, std::vector<std::string>{"IntermediatePosition"},verbose);
-
-        return sweep.vIntermediatePositionImages;
-    }
 
     template <class PointType>
-    std::vector<std::vector<typename SimpleXMLParser<PointType>::IntermediatePositionImages>> loadIntermediatePositionImagesFromMultipleSweeps(std::string folderPath, bool verbose=false)
-    {
-        std::vector<std::vector<typename SimpleXMLParser<PointType>::IntermediatePositionImages>> toRet;
-
-        SimpleSummaryParser summary_parser;
-        summary_parser.createSummaryXML(folderPath);
-        auto sweep_xmls = summary_parser.getRooms();
-
-        for (size_t i=0; i<sweep_xmls.size(); i++)
-        {
-            auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweep_xmls[i], std::vector<std::string>{"IntermediatePosition"},verbose);
-            toRet.push_back(sweep.vIntermediatePositionImages);
-        }
-
-        return toRet;
-    }
+    std::vector<typename SimpleXMLParser<PointType>::IntermediatePositionImages> loadIntermediatePositionImagesFromSingleSweep(std::string sweepXmlPath, bool verbose=false);
 
     template <class PointType>
-    std::vector<std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>>> loadIntermediatePositionImagesForTopologicalWaypoint(std::string folderPath, std::string waypoint,bool verbose=false)
-    {
-        std::vector<std::vector<typename SimpleXMLParser<PointType>::IntermediatePositionImages>> toRet;
+    std::vector<std::vector<typename SimpleXMLParser<PointType>::IntermediatePositionImages>> loadIntermediatePositionImagesFromMultipleSweeps(std::string folderPath, bool verbose=false);
 
-        SimpleSummaryParser summary_parser;
-        summary_parser.createSummaryXML(folderPath);
-        auto sweep_xmls = summary_parser.getRooms();
+    template <class PointType>
+    std::vector<std::vector<boost::shared_ptr<pcl::PointCloud<PointType>>>> loadIntermediatePositionImagesForTopologicalWaypoint(std::string folderPath, std::string waypoint,bool verbose=false);
 
-        // first construct waypoint id to sweep xml map
-        std::map<std::string, std::vector<std::string>> waypointToSweepsMap;
-        for (size_t i=0; i<sweep_xmls.size(); i++)
-        {
-            auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweep_xmls[i].roomXmlFile, std::vector<std::string>(), verbose);
-            waypointToSweepsMap[sweep.roomWaypointId].push_back(sweep_xmls[i].roomXmlFile);
-        }
-
-        auto sweepsAtWaypoint = waypointToSweepsMap[waypoint];
-
-        for (size_t i=0; i<sweepsAtWaypoint.size(); i++)
-        {
-            auto sweep = SimpleXMLParser<PointType>::loadRoomFromXML(sweepsAtWaypoint[i], std::vector<std::string>{"IntermediatePosition"},verbose);
-            toRet.push_back(sweep.vIntermediatePositionImages);
-        }
-
-        return toRet;
-    }
+#include "load_utilities.hpp"
 
 }
